Fix erroneous return value of gnc_gconf_get_float()
Discovered in bug#663433 - the return value was cached in an int variable
instead of a gdouble one.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@21541 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/app-utils/gnc-ui-util.c
@@ -2233,7 +2233,7 @@ gnc_auto_decimal_init (void)
     auto_decimal_enabled =
         gnc_gconf_get_bool(GCONF_GENERAL, "auto_decimal_point", NULL);
     auto_decimal_places =
-        gnc_gconf_get_float(GCONF_GENERAL, "auto_decimal_places", NULL);
+        (int)gnc_gconf_get_float(GCONF_GENERAL, "auto_decimal_places", NULL);
 }
 
 void
